Introduce a factory method for initializing the agent controller. Pull methods up from the Agent service to controller.
