+++ build.gradle
@@ -275,5 +275,5 @@ task allDependencies {
 }
 
 ext.versions = [
-  jetty: '9.2.18.v20160721'
+  jetty: '9.2.19.v20160908'
 ]
