+++ common/src/com/thoughtworks/go/domain/builder/FetchArtifactBuilder.java
@@ -1,18 +1,20 @@
-/*************************GO-LICENSE-START*********************************
+/*************************
+ * GO-LICENSE-START*********************************
  * Copyright 2016 ThoughtWorks, Inc.
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *************************GO-LICENSE-END***********************************/
+ * ************************GO-LICENSE-END
+ ***********************************/
 
 package com.thoughtworks.go.domain.builder;
 
@@ -23,12 +25,8 @@ import com.thoughtworks.go.util.URLService;
 import com.thoughtworks.go.util.command.CruiseControlException;
 import com.thoughtworks.go.util.command.EnvironmentVariableContext;
 import com.thoughtworks.go.work.DefaultGoPublisher;
-import org.apache.log4j.Logger;
 
 public class FetchArtifactBuilder extends Builder {
-    public static Logger LOG = Logger.getLogger(FetchArtifactBuilder.class);
-
-
     private final JobIdentifier jobIdentifier;
     private String srcdir;
     private final String dest;
@@ -95,9 +93,9 @@ public class FetchArtifactBuilder extends Builder {
 
     @Override
     public BuildCommand buildCommand() {
-            String checksumUrl = String.format("/remoting/files/%s/%s/%s", jobIdentifier.buildLocator(), ArtifactLogUtil.CRUISE_OUTPUT_FOLDER, ArtifactLogUtil.MD5_CHECKSUM_FILENAME);
-            return BuildCommand.compose(
-                    BuildCommand.echoWithPrefix(String.format("Fetching artifact [%s] from [%s]", getSrc(), jobLocatorForDisplay())),
-                    handler.toDownloadCommand(artifactLocator(), checksumUrl, checksumFileHandler.getChecksumFile()));
+        String checksumUrl = String.format("/remoting/files/%s/%s/%s", jobIdentifier.buildLocator(), ArtifactLogUtil.CRUISE_OUTPUT_FOLDER, ArtifactLogUtil.MD5_CHECKSUM_FILENAME);
+        return BuildCommand.compose(
+                BuildCommand.echoWithPrefix(String.format("Fetching artifact [%s] from [%s]", getSrc(), jobLocatorForDisplay())),
+                handler.toDownloadCommand(artifactLocator(), checksumUrl, checksumFileHandler.getChecksumFile()));
     }
 }
